requirement.txt
pycryptodome==3.15.0
PyJWT==2.4.0
Flask==2.0.3
redis==4.3.5
PyMySQL==1.0.2